Small changes based on PR feedback
