+++ packages/cli/test/unit/commands/pg/backups/index.unit.test.ts
@@ -3,7 +3,7 @@ import * as nock from 'nock'
 import {stdout} from 'stdout-stderr'
 import heredoc from 'tsheredoc'
 import Cmd from '../../../../../src/commands/pg/backups/index'
-import type {BackupTransfer} from '../../../../../src/lib/pg/backups'
+import type {BackupTransfer} from '../../../../../src/lib/pg/types'
 import runCommand from '../../../../helpers/runCommand'
 
 describe('pg:backups', () => {
